backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
azure-ai-formrecognizer==3.2.1
azure-cosmos==4.5.1
azure-search-documents==11.4.0
azure-identity==1.14.0
azure-storage-blob==12.17.0
autogen==0.2.0
openai==1.3.0
numpy==1.24.3
pandas==2.0.3
python-docx==0.8.11
pydantic==2.4.2
python-dotenv==1.0.0 